django_api/requirements.txt
boto3==1.37.18
Django==5.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
